Split ClawExec backend and tests into per-concern files
Decompose the two ClawExec monoliths into one-type/one-concern satellites
under Backend/, Workspace/, and Wire/ subdirectories, and split the
1000-line backend test suite into five concern files with shared fixtures
moved to ClawExecTestSupport.

Pure mechanical relocation: the only edits beyond moving code are dropping
`private` from extension blocks and type declarations now referenced across
the new file boundaries, and adding the imports each file needs. No logic,
signature, comment, string, or test assertion changed; normalized content
is identical to the originals and the ClawExec suite stays green.
