+++ Tests/ClawExecTests/ClawExecTestSupport.swift
@@ -1,4 +1,6 @@
+import ClawCore
 import Foundation
+import Testing
 
 @testable import ClawExec
 
@@ -128,5 +130,179 @@ func scratchChildren(_ stateRoot: URL) throws -> [URL] {
   return try FileManager.default.contentsOfDirectory(
     at: root,
     includingPropertiesForKeys: nil
+  )
+}
+
+struct ScratchFixture {
+  let root: URL
+
+  init() throws {
+    root = FileManager.default.temporaryDirectory
+      .appending(path: "clawd-scratch-tests-\(UUID().uuidString.lowercased())")
+    try FileManager.default.createDirectory(at: root, withIntermediateDirectories: false)
+  }
+
+  func remove() {
+    try? FileManager.default.removeItem(at: root)
+  }
+}
+
+func fixedIdentity() throws -> ExecutionIdentity {
+  ExecutionIdentity(uuid: try #require(UUID(uuidString: "11111111-2222-3333-4444-555555555555")))
+}
+
+func pythonEntrypoint() -> StagedFile {
+  StagedFile(name: ".clawd-entrypoint.py", bytes: Data("print('ok')".utf8), mode: .readExecute)
+}
+
+func executionRequest(input: StagedFile? = nil) -> ExecutionRequest {
+  ExecutionRequest(
+    language: .python,
+    entrypoint: pythonEntrypoint(),
+    inputs: input.map { staged in
+      [staged]
+    } ?? [],
+    network: false,
+    timeout: .seconds(1)
+  )
+}
+
+func permissions(_ url: URL) throws -> Int {
+  let attributes = try FileManager.default.attributesOfItem(atPath: url.path)
+  return try #require((attributes[.posixPermissions] as? NSNumber)?.intValue)
+}
+
+struct BackendFixture {
+  let root: URL
+  let settings: ExecSandboxSettings
+
+  init() throws {
+    root = FileManager.default.temporaryDirectory
+      .appending(path: "clawd-backend-tests-\(UUID().uuidString.lowercased())")
+    try FileManager.default.createDirectory(at: root, withIntermediateDirectories: false)
+    settings = ExecSandboxSettings(
+      workloadImage: try #require(
+        PinnedImageReference.parse(
+          "cgr.dev/swift-claw/python@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+        )
+      ),
+      memoryMiB: 1024,
+      cpus: 4
+    )
+  }
+
+  func backend(
+    commands: any ContainerCommandRunning = NoopCommandRunner(),
+    sanitizeReason: @escaping @Sendable (String) -> String = { $0 },
+    now: @escaping @Sendable () -> ContinuousClock.Instant = { ContinuousClock.now },
+    supportedHost: @escaping @Sendable () -> Bool = { true },
+    watchdogSleep: @escaping @Sendable (Duration) async throws -> Void = { duration in
+      try await Task.sleep(for: duration)
+    }
+  ) -> ContainerBackend {
+    ContainerBackend(
+      settings: settings,
+      stateRoot: root,
+      commands: commands,
+      sanitizeReason: sanitizeReason,
+      now: now,
+      supportedHost: supportedHost,
+      watchdogSleep: watchdogSleep
+    )
+  }
+
+  func remove() {
+    try? FileManager.default.removeItem(at: root)
+  }
+}
+
+// First call anchors the execution start; every later call sits far past the outer
+// deadline so the host watchdog fires deterministically without wall-clock waiting.
+final class SteppingNowSource: @unchecked Sendable {
+  private let lock = NSLock()
+  private let base = ContinuousClock.now
+  private var calls = 0
+
+  func next() -> ContinuousClock.Instant {
+    lock.lock()
+    defer { lock.unlock() }
+    calls += 1
+    return calls == 1 ? base : base.advanced(by: .seconds(3600))
+  }
+}
+
+struct NoopCommandRunner: ContainerCommandRunning {
+  func run(_ command: ContainerCommand) async -> ContainerCommandResult {
+    ContainerCommandResult(
+      termination: .exited(0),
+      stdout: CapturedCommandStream(bytes: Data(), totalBytes: 0, truncated: false),
+      stderr: CapturedCommandStream(bytes: Data(), totalBytes: 0, truncated: false),
+      processIdentifier: 1,
+      wallClock: .zero
+    )
+  }
+}
+
+actor AsyncGate {
+  private var isOpen = false
+  private var waiters: [CheckedContinuation<Void, Never>] = []
+
+  func wait() async {
+    if isOpen { return }
+    await withCheckedContinuation { continuation in
+      waiters.append(continuation)
+    }
+  }
+
+  func open() {
+    isOpen = true
+    let pending = waiters
+    waiters.removeAll()
+    for continuation in pending {
+      continuation.resume()
+    }
+  }
+}
+
+actor ExecutionRecorder {
+  private var recorded: [Int] = []
+  private var waiters: [(Int, CheckedContinuation<Void, Never>)] = []
+
+  func record(_ value: Int) {
+    recorded.append(value)
+    let ready = waiters.filter { waiter in
+      recorded.count >= waiter.0
+    }
+    waiters.removeAll { waiter in
+      recorded.count >= waiter.0
+    }
+    for waiter in ready {
+      waiter.1.resume()
+    }
+  }
+
+  func waitForCount(_ count: Int) async {
+    if recorded.count >= count { return }
+    await withCheckedContinuation { continuation in
+      waiters.append((count, continuation))
+    }
+  }
+
+  func values() -> [Int] { recorded }
+}
+
+func waitForQueuedCount(_ count: Int, backend: ContainerBackend) async {
+  while await backend.queuedExecutionCountForTesting < count {
+    await Task.yield()
+  }
+}
+
+func testExecutionResult(code: Int32) -> ExecutionResult {
+  ExecutionResult(
+    terminationReason: .exited(code: code),
+    stdout: "",
+    stderr: "",
+    truncatedRawBytes: false,
+    wallClock: .zero
   )
 }
